removes unecessary cast

+++ test/mapping-lite-document-test/src/main/java/org/eclipse/jnosql/lite/mapping/entities/ActorBuilder.java
@@ -31,7 +31,7 @@ public class ActorBuilder {
     }
 
     public ActorBuilder withId() {
-        this.id = (long) 12;
+        this.id = 12;
         return this;
     }
